+++ src/FrontendFramework/AbstractFrontendFramework.php
@@ -9,6 +9,8 @@
 namespace HeimrichHannot\TwigTemplatesBundle\FrontendFramework;
 
 use Contao\LayoutModel;
+use HeimrichHannot\TwigTemplatesBundle\Event\BeforeRenderCallback;
+use HeimrichHannot\TwigTemplatesBundle\Event\PrepareTemplateCallback;
 use HeimrichHannot\TwigTemplatesBundle\Twig\AbstractTemplate;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -34,6 +36,8 @@ abstract class AbstractFrontendFramework
     /**
      * Return the framework alias. Is used for template suffix and database identification.
      * Example: bs4 for Bootstrap 4.
+     *
+     * @deprecated use FrontendFrameworkInterface::getIdentifier instead
      */
     abstract public function getAlias(): string;
 
@@ -51,6 +55,28 @@ abstract class AbstractFrontendFramework
      * Update template data and template name before render template.
      */
     abstract public function compile(string &$templateName, array &$templateData, AbstractTemplate $entity): void;
+
+    public function prepare(PrepareTemplateCallback $callback): PrepareTemplateCallback
+    {
+        $templateName = $callback->getTemplateName();
+        $templateData = $callback->getData();
+        $this->generate($templateName, $templateData);
+        $callback->setData($templateData);
+
+        return $callback;
+    }
+
+    public function beforeRender(BeforeRenderCallback $callback): BeforeRenderCallback
+    {
+        $templateName = $callback->getTwigTemplateName();
+        $templateData = $callback->getTwigTemplateContext();
+        $entity = $callback->getLegacyTemplate();
+        $this->compile($templateName, $templateData, $entity);
+        $callback->setTwigTemplateName($templateName);
+        $callback->setTwigTemplateContext($templateData);
+
+        return $callback;
+    }
 
     protected function getLayout()
     {