+++ src/Resources/contao/config/config.php
@@ -1,14 +1,10 @@
 <?php
 
-/**
- * Hooks
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
  */
-$GLOBALS['TL_HOOKS']['parseTemplate']['twig-templates']    = ['huh.twig.event_listener.hook_listener', 'parseTemplate'];
 
-if (is_array($GLOBALS['TL_HOOKS']['parseWidget']) && in_array(['HeimrichHannot\FormHybrid\Hooks', 'parseWidgetHook'], $GLOBALS['TL_HOOKS']['parseWidget'])) {
-    array_insert($GLOBALS['TL_HOOKS']['parseWidget'], array_search(['HeimrichHannot\FormHybrid\Hooks', 'parseWidgetHook'], $GLOBALS['TL_HOOKS']['parseWidget']), [
-        'twig-templates' => ['huh.twig.event_listener.hook_listener', 'parseWidget']
-    ]);
-} else {
-    $GLOBALS['TL_HOOKS']['parseWidget']['twig-templates']    = ['huh.twig.event_listener.hook_listener', 'parseWidget'];
-}
+$GLOBALS['TL_HOOKS']['parseTemplate']['twig-templates'] = [\HeimrichHannot\TwigTemplatesBundle\EventListener\RenderListener::class, 'onParseTemplate'];
+$GLOBALS['TL_HOOKS']['parseWidget']['twig-templates'] = [\HeimrichHannot\TwigTemplatesBundle\EventListener\RenderListener::class, 'onParseWidget'];