+++ src/Twig/TemplateFactory.php
@@ -18,6 +18,11 @@ use HeimrichHannot\UtilsBundle\Template\TemplateUtil;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
+/**
+ * Class TemplateFactory.
+ *
+ * @deprecated
+ */
 class TemplateFactory
 {
     /**