+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.0-BETA-DEV] - 2020-08-04
+- added proxy template to replace html5 template dublicates
+- added FrontendFrameworkInterface to soon replace AbstractFrontendFramework
+- refactored HookListener into RenderListener
+- deprecated AbstractTemplate and its childs
+
 ## [1.3.0-DEV] - 2020-08-04
 - removed form_captcha template as it not used and working anymore
 - added blocks and formAttributes variable to member_default template